Keep server modules out of browser chunks
