+++ packages/next/src/client/components/error-boundary.tsx
@@ -4,7 +4,6 @@ import React, { type JSX } from 'react'
 import { useUntrackedPathname } from './navigation-untracked'
 import { isNextRouterError } from './is-next-router-error'
 import { handleHardNavError } from './nav-failure-handler'
-import { workAsyncStorage } from '../../server/app-render/work-async-storage.external'
 
 const styles = {
   error: {
@@ -54,10 +53,15 @@ interface ErrorBoundaryHandlerState {
 // function crashes so we can maintain our previous cache
 // instead of caching the error page
 function HandleISRError({ error }: { error: any }) {
-  const store = workAsyncStorage.getStore()
-  if (store?.isRevalidate || store?.isStaticGeneration) {
-    console.error(error)
-    throw error
+  if (typeof window === 'undefined') {
+    const { workAsyncStorage } =
+      require('../../server/app-render/work-async-storage.external') as typeof import('../../server/app-render/work-async-storage.external')
+
+    const store = workAsyncStorage.getStore()
+    if (store?.isRevalidate || store?.isStaticGeneration) {
+      console.error(error)
+      throw error
+    }
   }
 
   return null
