+++ packages/next/src/client/components/match-segments.ts
@@ -1,4 +1,3 @@
-import { getSegmentParam } from '../../server/app-render/get-segment-param'
 import type { Segment } from '../../server/app-render/types'
 
 export const matchSegment = (
@@ -18,18 +17,4 @@ export const matchSegment = (
     return false
   }
   return existingSegment[0] === segment[0] && existingSegment[1] === segment[1]
-}
-
-/*
- * This function is used to determine if an existing segment can be overridden by the incoming segment.
- */
-export const canSegmentBeOverridden = (
-  existingSegment: Segment,
-  segment: Segment
-): boolean => {
-  if (Array.isArray(existingSegment) || !Array.isArray(segment)) {
-    return false
-  }
-
-  return getSegmentParam(existingSegment)?.param === segment[0]
 }
