+++ packages/next/src/client/components/navigation.ts
@@ -15,7 +15,13 @@ import {
 import { getSegmentValue } from './router-reducer/reducers/get-segment-value'
 import { PAGE_SEGMENT_KEY, DEFAULT_SEGMENT_KEY } from '../../shared/lib/segment'
 import { ReadonlyURLSearchParams } from './navigation.react-server'
-import { useDynamicRouteParams } from '../../server/app-render/dynamic-rendering'
+
+const useDynamicRouteParams =
+  typeof window === 'undefined'
+    ? (
+        require('../../server/app-render/dynamic-rendering') as typeof import('../../server/app-render/dynamic-rendering')
+      ).useDynamicRouteParams
+    : undefined
 
 /**
  * A [Client Component](https://nextjs.org/docs/app/building-your-application/rendering/client-components) hook
@@ -84,7 +90,7 @@ export function useSearchParams(): ReadonlyURLSearchParams {
  */
 // Client components API
 export function usePathname(): string {
-  useDynamicRouteParams('usePathname()')
+  useDynamicRouteParams?.('usePathname()')
 
   // In the case where this is `null`, the compat types added in `next-env.d.ts`
   // will add a new overload that changes the return type to include `null`.
@@ -144,7 +150,7 @@ export function useRouter(): AppRouterInstance {
  */
 // Client components API
 export function useParams<T extends Params = Params>(): T {
-  useDynamicRouteParams('useParams()')
+  useDynamicRouteParams?.('useParams()')
 
   return useContext(PathParamsContext) as T
 }
@@ -215,7 +221,7 @@ function getSelectedLayoutSegmentPath(
 export function useSelectedLayoutSegments(
   parallelRouteKey: string = 'children'
 ): string[] {
-  useDynamicRouteParams('useSelectedLayoutSegments()')
+  useDynamicRouteParams?.('useSelectedLayoutSegments()')
 
   const context = useContext(LayoutRouterContext)
   // @ts-expect-error This only happens in `pages`. Type is overwritten in navigation.d.ts
@@ -246,7 +252,7 @@ export function useSelectedLayoutSegments(
 export function useSelectedLayoutSegment(
   parallelRouteKey: string = 'children'
 ): string | null {
-  useDynamicRouteParams('useSelectedLayoutSegment()')
+  useDynamicRouteParams?.('useSelectedLayoutSegment()')
 
   const selectedLayoutSegments = useSelectedLayoutSegments(parallelRouteKey)
