+++ packages/next/src/client/request/search-params.browser.dev.ts
@@ -1,29 +1,17 @@
-import type { SearchParams } from './search-params'
+import type { SearchParams } from '../../server/request/search-params'
 
-import { ReflectAdapter } from '../web/spec-extension/adapters/reflect'
+import { ReflectAdapter } from '../../server/web/spec-extension/adapters/reflect'
 import {
   describeStringPropertyAccess,
   describeHasCheckingStringProperty,
   wellKnownProperties,
-} from './utils'
-
-export function createRenderSearchParamsFromClient(
-  underlyingSearchParams: SearchParams
-): Promise<SearchParams> {
-  if (process.env.NODE_ENV === 'development') {
-    return makeUntrackedExoticSearchParamsWithDevWarnings(
-      underlyingSearchParams
-    )
-  } else {
-    return makeUntrackedExoticSearchParams(underlyingSearchParams)
-  }
-}
+} from '../../shared/lib/utils/reflect-utils'
 
 interface CacheLifetime {}
-const CachedSearchParams = new WeakMap<CacheLifetime, Promise<SearchParams>>()
 
-function makeUntrackedExoticSearchParamsWithDevWarnings(
-  underlyingSearchParams: SearchParams
+export function makeUntrackedExoticSearchParamsWithDevWarnings(
+  underlyingSearchParams: SearchParams,
+  CachedSearchParams: WeakMap<CacheLifetime, Promise<SearchParams>>
 ): Promise<SearchParams> {
   const cachedSearchParams = CachedSearchParams.get(underlyingSearchParams)
   if (cachedSearchParams) {
@@ -94,32 +82,6 @@ function makeUntrackedExoticSearchParamsWithDevWarnings(
 
   CachedSearchParams.set(underlyingSearchParams, proxiedPromise)
   return proxiedPromise
-}
-
-function makeUntrackedExoticSearchParams(
-  underlyingSearchParams: SearchParams
-): Promise<SearchParams> {
-  const cachedSearchParams = CachedSearchParams.get(underlyingSearchParams)
-  if (cachedSearchParams) {
-    return cachedSearchParams
-  }
-
-  // We don't use makeResolvedReactPromise here because searchParams
-  // supports copying with spread and we don't want to unnecessarily
-  // instrument the promise with spreadable properties of ReactPromise.
-  const promise = Promise.resolve(underlyingSearchParams)
-  CachedSearchParams.set(underlyingSearchParams, promise)
-
-  Object.keys(underlyingSearchParams).forEach((prop) => {
-    if (wellKnownProperties.has(prop)) {
-      // These properties cannot be shadowed because they need to be the
-      // true underlying value for Promises to work correctly at runtime
-    } else {
-      ;(promise as any)[prop] = underlyingSearchParams[prop]
-    }
-  })
-
-  return promise
 }
 
 function warnForSyncAccess(expression: string) {
