+++ packages/next/src/lib/metadata/metadata.tsx
@@ -3,7 +3,7 @@ import type { ParsedUrlQuery } from 'querystring'
 import type { GetDynamicParamFromSegment } from '../../server/app-render/app-render'
 import type { LoaderTree } from '../../server/lib/app-dir-module'
 import type { CreateServerParamsForMetadata } from '../../server/request/params'
-import type { StreamingMetadataResolvedState } from './async-metadata'
+import type { StreamingMetadataResolvedState } from '../../client/components/metadata/async-metadata'
 import {
   AppleWebAppMeta,
   FormatDetectionMeta,
@@ -37,7 +37,10 @@ import {
   METADATA_BOUNDARY_NAME,
   VIEWPORT_BOUNDARY_NAME,
 } from './metadata-constants'
-import { AsyncMetadata, AsyncMetadataOutlet } from './async-metadata'
+import {
+  AsyncMetadata,
+  AsyncMetadataOutlet,
+} from '../../client/components/metadata/async-metadata'
 import { isPostpone } from '../../server/lib/router-utils/is-postpone'
 
 // Use a promise to share the status of the metadata resolving,
