+++ packages/next/src/server/app-render/collect-segment-data.tsx
@@ -26,7 +26,7 @@ import {
   encodeSegment,
   ROOT_SEGMENT_KEY,
   type EncodedSegment,
-} from './segment-value-encoding'
+} from '../../shared/lib/segment-cache/segment-value-encoding'
 import { getDigestForWellKnownError } from './create-error-handler'
 import type { FallbackRouteParams } from '../request/fallback-params'
