+++ packages/next/src/shared/lib/router/utils/route-regex.ts
@@ -2,7 +2,7 @@ import {
   NEXT_INTERCEPTION_MARKER_PREFIX,
   NEXT_QUERY_PARAM_PREFIX,
 } from '../../../../lib/constants'
-import { INTERCEPTION_ROUTE_MARKERS } from '../../../../server/lib/interception-routes'
+import { INTERCEPTION_ROUTE_MARKERS } from './interception-routes'
 import { escapeStringRegexp } from '../../escape-regexp'
 import { removeTrailingSlash } from './remove-trailing-slash'
