+++ packages/next/src/shared/lib/segment-cache/segment-value-encoding.ts
@@ -1,5 +1,5 @@
-import { PAGE_SEGMENT_KEY } from '../../shared/lib/segment'
-import type { Segment as FlightRouterStateSegment } from './types'
+import { PAGE_SEGMENT_KEY } from '../segment'
+import type { Segment as FlightRouterStateSegment } from '../../../server/app-render/types'
 
 // TypeScript trick to simulate opaque types, like in Flow.
 type Opaque<K, T> = T & { __brand: K }
